requirements.txt
PyYAML~=6.0
PyMySQL~=1.0.2
Jinja2~=3.1.2